requirements.txt
tensorflow
matplotlib
pandas
seaborn
flask
opencv-python~=4.10.0.84
mediapipe~=0.10.14
scikit-learn~=1.5.2
numpy~=1.24.3
pyttsx3~=2.98
pillow~=10.4.0